Reorganised files.Code refactoring.
