+++ src/server/common/inferServerSideProps.ts
@@ -4,8 +4,5 @@ type GetSSRResult<TProps> =
 
 type GetSSRFn<TProps> = (...args: any[]) => Promise<GetSSRResult<TProps>>;
 
-export type ServerSideProps<TFn extends GetSSRFn<any>> = TFn extends GetSSRFn<
-  infer TProps
->
-  ? NonNullable<TProps>
-  : never;
+export type inferServerSideProps<TFn extends GetSSRFn<any>> =
+  TFn extends GetSSRFn<infer TProps> ? NonNullable<TProps> : never;
