+++ src/server/trpc/context.ts
@@ -2,7 +2,7 @@ import { type inferAsyncReturnType } from "@trpc/server";
 import { type CreateNextContextOptions } from "@trpc/server/adapters/next";
 import { type Session } from "next-auth";
 
-import { getServerAuthSession } from "../common/get-server-auth-session";
+import { getServerAuthSession } from "../common/getServerAuthSession";
 import { prisma } from "../db/client";
 
 type CreateContextOptions = {
